add last message to chat

+++ src/main/java/org/green/chat/model/MessageReadRequest.java
@@ -1,4 +1,4 @@
 package org.green.chat.model;
 
-public record MessageReadRequest(int chatId, int fromId, int toId) {
+public record MessageReadRequest(long chatId, int fromId, int toId) {
 }
